Actions performed:
Add - In the Buidling POCO - added the BuidlingGUID.
Fix - On the load buidlings. Changed the system to look at the GUId
instead of the ID (int).
Add - In the resources added 2 more strings for the buidlings import -
failure and file does not exist.
Fix - On the import file part - added check if the file exist or not.