Fixed the following:
Empty items on the edit and remove.
Edit able to edit items not on the page.
Zero item on the edit - fixed.

Updated the History Version file.

Created batch files to deploy to Prod, Dev, QA.

Added a batch file to copy code from local to backup.

Tweaked the rollback batch file.

Fixed issue on the simulate - on the date.