Worked on the following:
--------------------------------------------
Relocating the enumerations to "EnumLib.dll"
--------------------------------------------

The UniversalUtilities class , generated part of the ArchCorpUtilities.dll have a draw back.
When new items was created Methods, properties in the UniversalUtilities - it would dissapear after the codeGen.
Because it was not part of the generation logic only the enumerations was generated. Which consist of Navigation, MenuDomain, CodeTemplateEnum, MenuTypeEnum and CommandEnum.
Which is used in alot of places (over 20 places estimated.).

Intermediate solution - copy the changed files back to the backup folder with the cirrent GUID. Issue - it is a lot of work. Changing in multiple places.
Problem the rollback copies the backup item into place. If the changes are not there then pooF!!! all gone. Redo... Arghhhh. :(

(Light bulp!) Solution, created a new dll called "EnumLib.dll". Which only contains the enumeration mentioned.
Plan is to generate this code for the "EnumLib.dll". Re-mamped all the code not to use the UniverasalUtilities but now "EnumLib".
As "using E = EnumLib.EnumLib;". Replaced U.<Enum> with E.<Enum>.

Change the code gen code to generate this file. Add the file in to the rollback batch file. Change the generation code to now make the change in the "EnumLib"
instead of the "UniversalUtilities". Roadblock though - the working folder needs to move one level back :0. Thus update all the places where the this happen as well. :(.

Found 2 more enumerations:

CodeTemplateEnum, MenuTypeEnum.

[The buterfly flaps its wings!] And a tornado strikes on mars and the ring of jupiter shifts.